saleor/tests/requirements_dev.txt
# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Test data
factory-boy==3.3.0
Faker==19.6.1
freezegun==1.2.2

# GraphQL testing
graphene-django==3.1.5

# Mocking
responses==0.23.3

# Code quality (optional)
black==23.9.1
flake8==6.1.0